+++ src/php/Compiler/Infrastructure/GlobalEnvironmentSingleton.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Phel\Compiler\Infrastructure;
 
+use Phel;
 use Phel\Compiler\Domain\Analyzer\Environment\GlobalEnvironment;
 use Phel\Compiler\Domain\Analyzer\Environment\GlobalEnvironmentInterface;
 use Phel\Compiler\Domain\Analyzer\Exceptions\GlobalEnvironmentAlreadyInitializedException;
 use Phel\Compiler\Domain\Analyzer\Exceptions\GlobalEnvironmentNotInitializedException;
-use Phel\Lang\Registry;
 
 final class GlobalEnvironmentSingleton
 {
@@ -51,11 +51,11 @@ final class GlobalEnvironmentSingleton
     }
 
     /**
-     * @interal
+     * @internal
      */
     public static function initializeNew(): GlobalEnvironmentInterface
     {
-        Registry::getInstance()->clear();
+        Phel::clear();
         self::$instance = new GlobalEnvironment();
 
         return self::$instance;