+++ tests/php/Unit/Lang/RegistryTest.php
@@ -40,7 +40,7 @@ final class RegistryTest extends TestCase
 
     public function test_error_when_non_existing_definition_by_reference(): void
     {
-        $this->expectExceptionMessage('Only variables can be returned by reference');
+        $this->expectExceptionMessage('Definition "ns/non-existing" not found');
 
         $this->registry->getDefinitionReference('ns', 'non-existing');
     }